\name{cipsr}

\alias{cipsr}

\docType{package}

\title{
An R interface to the Organon and Cipsanon models
}

\description{
Predicts individual tree growth, yield, and wood quality attributes
}

\details{
\tabular{ll}{
	Package: \tab cipsr \cr
	Type: \tab Package \cr
	Version: \tab 2.1 \cr
	Date: \tab 2014-11-14 \cr
	License: \tab GPL > 2 \cr
}
}

\author{
Nathaniel Osborne \email{nathaniel.osborne@oregonstate.edu} [aut, cre] \cr
Douglas Maguire [aut] and David Hann [aut] \cr
}

\keyword{ package }

