
%%	$Id $	

\name{cipsr}
\alias{cipsr}

\title{Predicts forest growth, yield and wood quality using the ORGANON and CIPSANON models}

\description{
Returns a list of tree and stand information along with a list of found errors and warnings 
}

\usage{
cipsr( InputList )
}
		   
\arguments{
  \item{InputList}{a \code{\link{cipsrexam}} object.}
}

\details{

  The \code{cipsr} function returns a list of tree and stand information 
  along with a liast of found errors and warnings. The argument supplied
  should look like the \code{\link{sample.data}} object.
}


\references{

}

\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}} 
	   
\seealso{
  \code{\link{load.data}},
  \code{\link{cipsrexam}}
}

\examples{
}

\keyword{
}


