\name{get.template}

\alias{get.template}

\title{Creates an Excel template}

\description{
This function creates an Excel template in the current working directory.  The template can be 
renamed, cleared of data and supplied with new information from the user.  Information in the 
Excel template is identical to the list \code{\link{cipsrexam}}.  
}

\usage{
get.template()
}
		   
\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}} 
	   
\seealso{
	\code{\link{cipsrexam}}
	\code{\link{load.data}}
}

\keyword{
copy, paste, template 
}


