
%%	$Id $	

\name{grow}

\alias{grow}

\title{Predicts forest growth, yield and wood quality using the ORGANON and CIPSANON models}

\description{
Returns a list of tree and stand information along with a list of found errors and warnings 
}

\usage{
cipsr( InputList )
}
		   
\arguments{
}

\details{
}


\references{
}

\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}} 
	   
\seealso{
  \code{\link{load.data}},
}

\examples{
}

\keyword{
}


