
%%	$Id $	

\name{load.data}

\alias{load.data}

\title{Loads an Excel template for input data supplied to the ORGANON and CIPSANON forest growth models}

\description{
Returns a list of sample, unit and silvicultural prescriptions
}

\usage{
load.data( InputFile )
}
		   
\arguments{
}

\details{
}

\references{
}

\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}} 
	   
\seealso{
  \code{\link{cipsrexam}}
}

\examples{
}

\keyword{
}


