\name{cipsr-package}
\alias{cipsr-package}
\alias{cipsr}
\docType{package}
\title{
cipsr provides an R interface to the Cipsanon and Organon forest models
}
\description{
Allows the prediction of individual tree growth, yield and wood quality 
}
\details{
\tabular{ll}{
	Package: \tab cipsr\cr
	Type: \tab Package\cr
	Version: \tab 2.0\cr
	Date: \tab 2014-10-10\cr
	License: \tab GPL > 2\cr
	}
}
\author{
Nathaniel Osborne, Doug Maguire, David Hann
Maintainer: Nathaniel L. Osborne <Nathaniel.Osborne@oregonstate.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
