
%%	$Id $	

\name{load.data}
\alias{load.data}

\title{Loads an Excel template for input data supplied to the ORGANON and CIPSANON forest growth models}

\description{
Returns a list of sample, unit and silvicultural prescriptions
}

\usage{
load.data( InputFile )
}
		   
\arguments{
  \item{InputFile}{a \code{\link{cipsrexam}} object.}
}

\details{

  The \code{load.data} function returns a list of sample, unit and 
  silvicultural activities supplied in an Excel template. The returned
  list will look like the \code{\link{sample.data}} object.
}


\references{

}

\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}} 
	   
\seealso{
  \code{\link{cipsr}},
  \code{\link{cipsrexam}}
}

\examples{
}

\keyword{
}


